+++ libraries/microsoft-agents-storage-cosmos/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "microsoft-agents-cosmos"
+name = "microsoft-agents-storage-cosmos"
 version = "0.0.0a1"
 description = "A Cosmos DB storage library for Microsoft Agents"
 authors = [{name = "Microsoft Corporation"}]
@@ -14,10 +14,10 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "microsoft.agents.storage",
+    "microsoft-agents-hosting-core",
     "azure-core",
     "azure-cosmos",
 ]
 
 [project.urls]
-"Homepage" = "https://github.com/microsoft/microsoft-agents-protocol"
+"Homepage" = "https://github.com/microsoft/Agents"
